Homework 2 Completed
Gemini's in line code-completion was used for creating some variables in my thisViewModel class, there are comments stating where this took place. I turned code-completion off afterwards as I had forgotten that it was on. Gemini was also used to determine why Text("...") wouldn't work (It needed an item {} wrapper)